Refactor: to move prompt.py to writingnode/prompts
