+++ src/researchgraph/nodes/writingnode/prompts/writeup_prompt.py
